General re-shuffling of input and result files for tests.
The original yaml-based test cases are now under `tests/inputs/spec`
and everything else will be in sibling directories of it.

Similarly the local results (parse structures) associated with the
yaml-based tests are now under `tests/results/spec`, and everything
else is in sibling directories of it.

The `z_` prefix for the local results is now gone as part of that.

Started testsuite for the rendering backend.
Untested: partials and all forms of sections.